requirements-dev.txt
# HL7 Community Digest - Development Dependencies

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
responses>=0.24.0  # Mock HTTP requests

# Code quality
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0

# Type stubs
types-requests>=2.31.0
types-beautifulsoup4>=4.12.0
types-python-dateutil>=2.8.0
boto3-stubs[ses,s3,secretsmanager]>=1.34.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0
